requirements.txt
# Core ML and NLP libraries
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
sentence-transformers==2.2.2
faiss-cpu==1.7.4
torch==2.1.0
transformers==4.35.0

# Web framework and API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Data processing
pandas==2.1.3
numpy==1.24.3
jsonlines==3.1.0

# HuggingFace datasets
datasets==2.14.5
mlcroissant==0.0.1

# HTTP client for Ollama
httpx==0.25.2
requests==2.31.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
click==8.1.7

# Optional: GPU support (uncomment if using GPU)
# faiss-gpu==1.7.4
# torch==2.1.0+cu118 --index-url https://download.pytorch.org/whl/cu118 